% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-list.R
\name{as.list.lazylist}
\alias{as.list.lazylist}
\title{Export a portion of a stream as a list}
\usage{
\method{as.list}{lazylist}(x, ..., from, to)
}
\arguments{
\item{x}{A stream}

\item{...}{Not used}

\item{from}{Starting index}

\item{to}{Ending index}
}
\description{
Export a portion of a stream as a list
}
