% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_weighted.R
\name{merge_weighted}
\alias{merge_weighted}
\title{Merge two streams so that the result is ordered}
\usage{
merge_weighted(stream1, stream2, weight)
}
\arguments{
\item{stream1}{a lazy list}

\item{stream2}{a lazy list}

\item{weight}{a function with one required argument}
}
\description{
The weight function is used to order the elements
}
