% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazylist.R
\name{shift_left}
\alias{shift_left}
\title{Shift a stream to the left}
\usage{
shift_left(stream)
}
\arguments{
\item{stream}{A stream}
}
\description{
When viewing streams as sequences, \code{stream_cdr} is equivalent to 
left-shifting the sequence. This function is just for convenience and readability
}
