% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_at.R
\name{stream_at}
\alias{stream_at}
\title{Select elements of a stream}
\usage{
stream_at(stream, filter)
}
\arguments{
\item{stream}{A stream to select from}

\item{filter}{A stream of TRUE/FALSE values}
}
\description{
Select elements of a stream
}

