% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazylist.R
\name{stream_car}
\alias{stream_car}
\alias{stream_cdr}
\title{Stream accessors}
\usage{
stream_car(stream)

stream_cdr(stream)
}
\arguments{
\item{stream}{A stream}
}
\description{
\code{stream_car} returns the first element of a stream, and 
\code{stream_cdr} returns the rest of the stream
}
