% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{stream_filter}
\alias{stream_filter}
\title{Filter a stream using a predicate function}
\usage{
stream_filter(stream, predicate)
}
\arguments{
\item{stream}{The stream to be filtered}

\item{predicate}{A function that takes a stream element as an argument, and returns TRUE/FALSE}
}
\description{
Filter a stream using a predicate function
}
