% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_each.R
\name{stream_for_each}
\alias{stream_for_each}
\title{Perform a function on each element of a stream (without returning values)}
\usage{
stream_for_each(stream, fun, until = 5)
}
\arguments{
\item{stream}{A stream}

\item{fun}{A function to perform}

\item{until}{How many iterations}
}
\description{
Perform a function on each element of a stream (without returning values)
}
\details{
This function is only useful for the side effects of \code{fun}
}
