% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{stream_map}
\alias{stream_map}
\title{Map a function onto a stream or streams}
\usage{
stream_map(s, f, ...)
}
\arguments{
\item{s}{A single stream, or a list of streams, to be used as function arguments}

\item{f}{A function that takes the appropriate number of arguments}
}
\description{
Map a function onto a stream or streams
}
