% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.R
\name{stream_which}
\alias{stream_which}
\title{Which stream indices are TRUE?}
\usage{
stream_which(stream, predicate, start_from = 1)
}
\arguments{
\item{stream}{A stream}

\item{predicate}{A function that returns TRUE or FALSE}

\item{start_from}{Find indexes on or after this position}
}
\description{
Which stream indices are TRUE?
}
