% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_at.R
\name{subset}
\alias{subset}
\alias{stream_at}
\alias{stream_at_index}
\title{Select elements of a stream}
\usage{
stream_at(stream, filter)

stream_at_index(stream, index)
}
\arguments{
\item{stream}{A stream to select from}

\item{filter}{A stream of TRUE/FALSE values}

\item{index}{A stream of integer indexes}
}
\description{
Select elements of a stream
}
