% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_pairs.R
\name{weighted_pairs}
\alias{weighted_pairs}
\title{Create a stream of (ordered) pairs, from two individual streams}
\usage{
weighted_pairs(s1, s2, weight)
}
\arguments{
\item{s1}{A stream}

\item{s2}{A stream}

\item{weight}{a weight function that is used to order the resulting stream}
}
\description{
The output stream will be ordered in ascending order of the weight function on 
each pair.
}

