% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_ref.R
\name{stream_ref}
\alias{stream_ref}
\alias{[.lazylist}
\title{Get a single element of a stream}
\usage{
stream_ref(stream, n)

\method{[}{lazylist}(stream, n)
}
\arguments{
\item{stream}{A stream}

\item{n}{An integer}
}
\description{
Get a single element of a stream
}
